demo/semantic_segmentation/requirements.txt
ipykernel==6.13.0
lightning-flash[image]==0.8.1.post0
torch==1.13.1
# pytorch-lightning==1.4.9
mlflow==2.1.1
boto3==1.26.79
onnx==1.8.1
onnxruntime==1.14.0
click==8.0.4